Don't create an AX extension object unless necessary
This is a workaround for a bug in the ruby-openid gem which outputs a
"mode" parameter even if there are no other values being queried. This
causes some providers (such as Google) to prompt the user differently
(for example, Google says "site X is requesting information from your
Google account" instead of just "sign in to site X with your Google
account").